+++ src/components/modals/AdminModal.jsx
@@ -104,19 +104,19 @@ export default function AdminModal({ children }) {
           <Button variant="outlined" size="small" color="primary" component={Link} to="/dashboard">
             <HomeIcon />
           </Button>
-          <Button variant="contained" size="small" color="primary" startIcon={<PlaceIcon />} component={Link} to="/admin/locations">
+          <Button variant="contained" size="small" startIcon={<PlaceIcon />} component={Link} to="/admin/locations" color={location.pathname.includes("/admin/locations") ? "secondary" : "primary"}>
             Locations
           </Button>
-          <Button variant="contained" size="small" color="primary" startIcon={<EventIcon />} component={Link} to="/admin/events">
+          <Button variant="contained" size="small" startIcon={<EventIcon />} component={Link} to="/admin/events" color={location.pathname.includes("/admin/events") ? "secondary" : "primary"}>
             Events
           </Button>
-          <Button variant="contained" size="small" color="primary" startIcon={<DescriptionIcon />} component={Link} to="/admin/pages">
+          <Button variant="contained" size="small" startIcon={<DescriptionIcon />} component={Link} to="/admin/pages" color={location.pathname.includes("/admin/pages") ? "secondary" : "primary"}>
            Pages
           </Button>
-          <Button variant="contained" size="small" color="primary" startIcon={<ListIcon />}component={Link} to="/admin/lists">
+          <Button variant="contained" size="small" startIcon={<ListIcon />} component={Link} to="/admin/lists" color={location.pathname.includes("/admin/lists") ? "secondary" : "primary"}>
             Lists
           </Button>
-          <Button variant="contained" size="small" color="primary" component={Link} to="/admin/settings">
+          <Button variant="contained" size="small" component={Link} to="/admin/settings" color={location.pathname.includes("/admin/settings") ? "secondary" : "primary"}>
             <SettingsIcon />
           </Button>
         </Box>