+++ CHANGELOG.md
@@ -3,16 +3,19 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [2.3.0] - 2024-05-30 (beta)
+## [2.3.0] - 2024-06-03 (beta)
 
 Pages:
   - Add: support for Gallery images
+
+Admin:
+  - Add: Indicate the current admin section (colour selected navigation)
 
 Admin / Pages:
   - Add: insert from gallery
 
 App:
-  - Add: admi gallery component
+  - Add: admin gallery component
 
 ## [2.2.3] - 2024-05-29 (beta)